Refactor: move email and telephone validation to practitioner service
