+++ service/practitioner_service_test.go
@@ -1,4 +1,4 @@
-package utils
+package service
 
 import (
 	"testing"
